+++ module-domain-order/src/main/java/domain/order/Order.java
@@ -1,7 +1,6 @@
-package kitchenpos.order.domain;
+package domain.order;
 
 import common.entity.BaseIdEntity;
-import kitchenpos.order.application.OrderValidationException;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
