+++ module-domain-order/src/main/java/domain/order/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package domain.order;
 
 import common.entity.BaseSeqEntity;
 import common.entity.Quantity;
