+++ module-domain-order/src/main/java/domain/order/OrderRepository.java
@@ -1,9 +1,7 @@
-package kitchenpos.order.domain;
+package domain.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
