+++ module-domain-order/src/main/java/domain/order/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package domain.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
