+++ module-domain-order/src/test/java/domain/order/OrderTableTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package domain.order;
 
 import common.entity.NumberOfGuests;
 import common.entity.Quantity;
@@ -6,7 +6,6 @@ import domain.menu.Menu;
 import domain.menu.MenuGroup;
 import domain.menu.MenuProduct;
 import domain.menu.Product;
-import kitchenpos.order.application.TableValidationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
