+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.dto;
 
-import kitchenpos.order.domain.OrderLineItem;
+import domain.order.OrderLineItem;
 
 public class OrderLineItemResponse {
