+++ src/main/java/kitchenpos/order/dto/OrderRequest_Create.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.dto;
 
 import domain.menu.Menu;
-import kitchenpos.order.domain.OrderItem;
+import domain.order.OrderItem;
 
 import java.util.List;
 import java.util.stream.Collectors;
